clangtidy

+++ csrc/ops/utils.h
@@ -37,7 +37,7 @@ IterType promoteIterType(IterType type1, IterType type2);
 std::vector<IterDomain*> mapMatmulOpIterDomains(
     const std::vector<IterDomain*>& input_domain,
     bool is_lhs,
-    int64_t out_size);
+    size_t out_size);
 
 IterDomain* newOutputIterDomain(const std::vector<IterDomain*>& ids);
